Mariam_Nousseir_part
This project implements a singly linked list in C++ with core functionalities including insertion at the end, checking if the list is empty or full, retrieving data by index, searching for a value, removing duplicate elements, and converting the list to a circular structure. Memory is managed carefully through a custom destructor that handles both linear and circular forms. Each function is designed to demonstrate safe pointer handling and dynamic memory allocation while ensuring correct list behavior through a series of tests in the main program.